Can configure a big_sync so that BIS1 is  RX and all the rest are TX. But not sending anything yet because notENABLE_SEND_ON_BIS2 is defined in main.c If the BIS2 is actaully sending something then it has fatal exception and the likelyhood is that some thread stack size is too small. Directly changing stack size is not possible as the west generate phase fails stating that those stack sizes are calculated and not settable.
Added the ability to compile some stack functions with no optimisation by prepending the function with the following code:-
#if defined(CONFIG_BT_ISO_BIS_RECV_SEND_DBG)
__attribute__((optimize("O0")))
#endif

When this firmware is flashed the log output is as follows which is weird but at least it is not crashing:-

S< 1:[1,1,722400,12]
S< 1:[-,-,-,13]
S< 1:[1,1,722425,12]
S< 1:[-,-,-,13]
S< 1:[1,1,722450,12]
S< 1:[-,-,-,13]
S< 1:[1,1,722475,12]
S< 1:[-,-,-,13]
S< 1:[1,1,722500,12]
S< 1:[-,-,-,13]
S< 1:[1,1,722525,12]

(*) Preparing for sending on BIS2 with the sync stream.

(*) Reverted hci.c so that sending iso will fail for send_recv in preperation for rework

(*) Refactor config names, consolidate, conditional checks on config.

(*) Implement capability to send on a BIS in a receiver

(*) Added the config CONFIG_BT_LL_SW_SPLIT=y to the iso_recv_send & iso_receive application

(*) Added config to build iso_recv_send with debug capability. But there is still something wrong

(*) Added capability to configure iso_recv_send for sending BIS2 or receiving BIS2 based on the #define ENABLE_SEND_ON_BIS2

It does not work and the app exits main with "failed (err -22)
Updated iso_recv_send so that it builds with ENABLE_SEND_ON_BIS2 in iso_recv_send/main.c either defined or not. Tested that without it defined I see the expected output in the log.
Note: Build with ENABLE_SEND_ON_BIS2 has not been tested.

- Initial minimal changes to customise from iso_receive.
- Including a big comment at the top describing how to incorporate latest upstream:main
- Config iso_broadcast so that it never terminates the BIG